Seperating config, changing structure, preparing Insert and Update till the last function (requestUpdate and requestInsert), seperating database connect to seperate JS.
