+++ Node/config.json
@@ -1,0 +1,7 @@
+{
+    "DBusername" : "sql7295537",
+    "DBpassword" : "8mI76qvWm6",
+    "DBhost" : "sql7.freesqldatabase.com",
+    "DBdatabase" : "sql7295537",
+    "Port" : "8081"
+}
